Function onionsalt::crypto::sillybox

pub fn sillybox(c: &mut [u8], m: &[u8], nauth: usize, n: &Nonce, pk: &PublicKey, sk: &SecretKey)

An implementation of public-key encryption similar to the NaCl function crypto_box (renamed crypto::sillybox_up in this package), but with the feature that it only authenticates the first nauth bytes. This is not useful for most purposes (thus its silly name), but is helpful for enabling round-trip onion routing in which all the routing information is authenticated (to information leaks triggered by maliciously modified packets), but information may be added to the communication en-route.