Function onionsalt::onionbox_open

pub fn onionbox_open(input: &[u8; PACKET_LENGTH], secret_key: &SecretKey) -> Result<OpenedOnionBox, NaClError>

The message is passed in input, and a struct is returned which has methods to access the decrypted routing information, the decrypted message (to be passed to the next router), to access the payload, and to insert an encrypted response to the payload.