Struct onionsalt::OnionBox

pub struct OnionBox {
    // some fields omitted
}

Methods

impl OnionBox

fn packet(&self) -> [u8; PACKET_LENGTH]

The encrypted packet, to be sent to the first receiver.

fn return_magic(&self) -> [u8; 32]

The response when we get it will begin with these bytes.

fn read_return(&self, payload_key: KeyPair, msg: &[u8; PACKET_LENGTH]) -> Result<[u8; PAYLOAD_LENGTH], NaClError>

This function accepts a packet that has been sent to us, and decrypts it without authentication if it is the response to our original message.

fn add_payload(&mut self, payload_key: KeyPair, payload_contents: &[u8; PAYLOAD_LENGTH]) -> &mut Self

fn add_encryptedpayload(&mut self, ciphertext: &[u8; ENCRYPTEDPAYLOAD_LENGTH]) -> &mut Self