Struct onionsalt::OpenedOnionBox

pub struct OpenedOnionBox {
    // some fields omitted
}

Methods

impl OpenedOnionBox

fn packet(&self) -> [u8; PACKET_LENGTH]

The packet to be forwarded onwards to the next router.

fn routing(&self) -> [u8; ROUTING_LENGTH]

The routing information for us.

fn key(&self) -> PublicKey

The payload public key of the sender (if we are the recipient).

fn payload(&self, response_key: &KeyPair) -> Result<[u8; PAYLOAD_LENGTH], NaClError>

Attempt to decrypt and authenticate the payload.

fn respond(&mut self, response_key: &KeyPair, response: &[u8; PAYLOAD_LENGTH])

Set response to the response payload information. This is only likely to work correctly if the sender intended to ask us for a response.